Move query client, fix typo, rename imports
