+++ webapp/src/main/frontend-v2/src/public/locales/fr-FR/branch.json
@@ -30,7 +30,7 @@
   "screenshotUploadModal.submit": "Importer",
   "screenshotUploadModal.noTextUnitsWithoutScreenshot": "Aucun texte sans capture d'écran",
   "screenshotUploadModal.screenshotDragUploadText": "Glissez et déposez les fichiers pour les importer",
-  "schreenshotUploadFailed": "Échec de l'importation des captures d'écran",
+  "screenshotUploadFailed": "Échec de l'importation des captures d'écran",
   "branches.search": "Rechercher par nom de branche ou auteur",
   "branches.name": "Nom",
   "branches.createdDate": "Date de création",
